+++ Strings and Arrays.cpp
@@ -1,14 +1,6 @@
 #include <iostream>
 #include <cctype>
-
-const int arraySizeTask1 = 10;
-const int arraySize = 255;
-
-void translateArray(int numbers[], int size);
-void toUpperCase(char str[]);
-bool isPalindrom(const char str[]);
-void parseStringLetters(const char str[], int& vowelsCount, int& consonantsCount);
-//bool isEqual(const char str1[], const char str2[]);
+#include "functionsHW9.h"
 
 int main()
 {
@@ -68,75 +60,18 @@ int main()
     // --------------- TASK 5 ---------------
     std::cout << "--------------- TASK 5 ---------------" << std::endl << std::endl;
 
+    char str1[arraySize]{};
+    char str2[arraySize]{};
 
-}
+    std::cout << "Enter character string 1: "; std::cin >> str1;
+    std::cout << "Enter character string 2: "; std::cin >> str2;
 
-void translateArray(int numbers[], int size) {
-    for (int i = 0; i < size; i++) {
-        if (numbers[i] > 0) {
-            std::cout << numbers[i] * 2 << " ";
-        }
-        else {
-            std::cout << 0 << " ";
-        }
+    if (isEqual(str1, str2)) {
+        std::cout << str1 << " is equal to " << str2;
     }
-}
-
-void toUpperCase(char str[]) {
-    for (int i = 0; i < strlen(str); i++) {
-        if (islower(str[i])) {
-            std::cout << static_cast<char>(std::toupper(str[i]));
-        }
-        else {
-            std::cout << static_cast<char>(str[i]);
-        }
+    else {
+        std::cout << str1 << " is not equal to " << str2;
     }
+
+    std::cout << std::endl;
 }
-
-bool isPalindrom(const char str[]) {
-    int wordLength = -1; 
-    for (int j = 0; j <= strlen(str); j++) {
-        if (str[j] == ' ' || str[j] == '\0') {
-            break;
-        }
-        else {
-            wordLength++;
-        }
-    }
-    
-    for (int i = 0; i <= wordLength - i; i++) {
-        if (str[i] != str[wordLength - i]) {
-            return false;
-            break;
-        }
-    }
-    return true;
-}
-
-void parseStringLetters(const char str[], int& vowelsCount, int& consonantsCount) {
-    int i = 0;
-
-    while (str[i] != ' ' && str[i] != '\0') {
-        bool isAlpha = std::isalpha(str[i]);
-        bool isVowel = std::tolower(str[i]) == 'a' ||
-            std::tolower(str[i]) == 'e' ||
-            std::tolower(str[i]) == 'o' ||
-            std::tolower(str[i]) == 'u' ||
-            std::tolower(str[i]) == 'i';
-
-        if (isAlpha && isVowel) {
-            vowelsCount++;
-        }
-        else if (isAlpha && !isVowel){
-            consonantsCount++;
-        }
-
-        i++;
-    }
-}
-
-//bool isEqual(const char str1[], const char str2[]) {
-//    if (strlen(str1) == strlen(str2)) {
-//        
-//    }
-//}
